refactor content blocks into admin namespace
