+++ app/controllers/admin/content_assets_controller.rb
