+++ app/views/admin/content_blocks/index.html.erb
@@ -1,7 +1,7 @@
 <h1 class="govuk-heading-l">Blocks</h1>
 <p class="govuk-body-s">A block is part of one of the fixed pages that you can edit. Blocks will usually be created by
   developers, only the content will be editable</p>
-<%= link_to 'Create a block', new_content_block_path, :class => "govuk-button" %>
+<%= link_to 'Create a block', new_admin_content_block_path, :class => "govuk-button" %>
 
 <div>
   <% if @content_blocks.count == 0 %>
@@ -17,7 +17,7 @@
         <p class="govuk-body-s">
           <%= content_block.description %>
           <span>
-            <%= link_to 'Edit', edit_content_block_path(content_block), { :class => "govuk-link" } %>
+            <%= link_to 'Edit', edit_admin_content_block_path(content_block), { :class => "govuk-link" } %>
           </span>
         </p>
       </div>
