+++ app/views/admin/content_blocks/new.html.erb
@@ -1,0 +1,8 @@
+<h1 class="govuk-heading-l">Add a new block</h1>
+
+<%= link_to 'Back', admin_content_blocks_path, class: 'govuk-back-link'%>
+
+<%= render 'form', content_block: @content_block %>
+
+<%= link_to 'Back', admin_content_blocks_path, class: 'govuk-back-link'%>
+
